# swagger/swagger.yaml
consumes:
    - application/json
definitions:
...
    BaseUserInfo:
        properties:
            Email:
                description: 用户邮箱
                type: string
            Telephone:
                description: 用户电话
                type: string
            UserId:
                description: 用户id
                type: string
            Username:
                description: 用户名称
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
    ContactsInfo:
        properties:
            CreatedAt:
                format: int64
                type: integer
            Id:
                format: int64
                type: integer
            Name:
                type: string
            UpdatedAt:
                format: int64
                type: integer
            Users:
                items:
                    $ref: '#/definitions/User'
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
    CreateContactRequest:
        description: CreateContactRequest 创建联系人组的请求数据
        properties:
            Name:
                description: 联系人组的名称
                type: string
            Users:
                description: 添加到联系人组的用户列表
                items:
                    $ref: '#/definitions/BaseUserInfo'
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
    CreateContactResponse:
        description: CreateContactResponse 创建联系人组的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
    DeleteContactRequest:
        description: DeleteContactRequest 删除联系人组的请求数据结构
        properties:
            Id:
                description: 要删除的联系人组的id
                format: int64
                type: integer
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
    DeleteContactResponse:
        description: DeleteContactResponse 删除联系人组的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
    ErrResponse:
        description: ErrResponse 错误响应
        properties:
            Code:
                example: 1001
                format: int64
                type: integer
            Data:
                example: {}
                type: object
            Message:
                example: 失败
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/alarm/v2/dto
...
    ListContactRequest:
        description: ListContactRequest 获取联系人组列表的请求数据结构
        properties:
            PageNo:
                description: 当前页码，从1开始
                format: int64
                type: integer
            PageSize:
                description: 每页数据个数，默认10个, 当值为-1时不分页
                format: int64
                type: integer
            SearchKey:
                description: 模糊搜索的key，允许值：name
                type: string
            SearchValue:
                description: 模糊搜索的value
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
    ListContactResponse:
        description: ListContactResponse 获取联系人组列表的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                $ref: '#/definitions/ListContactResponseData'
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
    ListContactResponseData:
        properties:
            Items:
                items:
                    $ref: '#/definitions/ContactsInfo'
                type: array
            Total:
                format: int64
                type: integer
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
    RemoveUserFromContactRequest:
        description: RemoveUserFromContactRequest 从联系人组中删除用户的请求数据结构
        properties:
            Id:
                format: int64
                type: integer
            UserId:
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
    RemoveUserFromContactResponse:
        description: RemoveUserFromContactResponse 从联系人组中删除用户的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
    UpdateContactRequest:
        description: UpdateContactRequest 编辑联系人组的请求数据结构
        properties:
            Id:
                description: 正在编辑的联系人组id
                format: int64
                type: integer
            Name:
                description: 联系人组名称
                type: string
            Users:
                description: 编辑后要加入联系人组的用户列表
                items:
                    $ref: '#/definitions/BaseUserInfo'
                type: array
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
    UpdateContactResponse:
        description: UpdateContactResponse 编辑联系人组的响应数据结构
        properties:
            Code:
                description: 响应码，0表示请求成功
                example: 0
                format: int64
                type: integer
            Data:
                type: object
            Message:
                example: 成功
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
    User:
        properties:
            CreatedAt:
                description: 加入用户组的时间
                format: int64
                type: integer
            Email:
                description: 用户邮箱
                type: string
            Telephone:
                description: 用户电话
                type: string
            UserId:
                description: 用户id
                type: string
            Username:
                description: 用户名称
                type: string
        type: object
        x-go-package: ezone.xxxxx.com/xxxxx/xxxxx/alarm/modules/contacts/dto
...
host: xxxxx.swagger.galaxy.cloud
info:
    description: 告警策略
    title: 告警策略
    version: 1.0.0
paths:
...
    /contact/create:
        post:
            description: 创建联系人组
            operationId: createContactRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/CreateContactRequest'
            responses:
                "200":
                    $ref: '#/responses/createContactResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 创建联系人组
            tags:
                - contact
    /contact/delete:
        post:
            description: 删除联系人组
            operationId: deleteContactRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/DeleteContactRequest'
            responses:
                "200":
                    $ref: '#/responses/deleteContactResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 删除联系人组
            tags:
                - contact
    /contact/list:
        post:
            description: 获取联系人组列表
            operationId: listContactRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/ListContactRequest'
            responses:
                "200":
                    $ref: '#/responses/listContactResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 获取联系人组列表
            tags:
                - contact
    /contact/remove:
        post:
            description: 从联系人组中移除用户
            operationId: removeUserFromContactRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/RemoveUserFromContactRequest'
            responses:
                "200":
                    $ref: '#/responses/removeUserFromContactResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 从联系人组中移除用户
            tags:
                - contact
    /contact/update:
        post:
            description: 编辑联系人组，同时更新联系人组中的联系人信息
            operationId: updateContactRequest
            parameters:
                - in: query
                  name: Action
                  type: string
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateContactRequest'
            responses:
                "200":
                    $ref: '#/responses/updateContactResponse'
                default:
                    $ref: '#/responses/errResponse'
            summary: 编辑联系人组
            tags:
                - contact
produces:
    - application/json
responses:
...
    createContactResponse:
        description: ""
        schema:
            $ref: '#/definitions/CreateContactResponse'
...
    deleteContactResponse:
        description: ""
        schema:
            $ref: '#/definitions/DeleteContactResponse'
...
    errResponse:
        description: 发生错误时的返回信息
        schema:
            $ref: '#/definitions/ErrResponse'
...
    listContactResponse:
        description: ""
        schema:
            $ref: '#/definitions/ListContactResponse'
...
    removeUserFromContactResponse:
        description: ""
        schema:
            $ref: '#/definitions/RemoveUserFromContactResponse'
...
    updateContactResponse:
        description: ""
        schema:
            $ref: '#/definitions/UpdateContactResponse'
...
schemes:
    - http
security:
    - api_key: []
securityDefinitions:
    api_key:
        in: header
        name: api_key
        type: apiKey
    basic:
        type: basic
swagger: "2.0"